\name{MKomics-package}
\alias{MKomics-package}
\alias{MKomics}
\docType{package}
\title{
Functions for Omics Data Analysis.
}
\description{
Contains several functions for analysis of omics data.
}
\details{
\tabular{ll}{
Package: \tab MKomics\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2020-01-06\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, utils, graphics, grDevices, RColorBrewer, robustbase, limma\cr
Suggests: \tab gplots, knitr, rmarkdown\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKomics)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
